Time: 3 ms (43.84%) | Memory: 8.4 MB (99.77%) - LeetSync

+++ 9-palindrome-number/palindrome-number.cpp
@@ -1,15 +1,15 @@
 class Solution {
 public:
-    bool isPalindrome(long long int x) {
+    bool isPalindrome(long int x) {
         if (x < 0) {
             return false;
         }
 
-        long long int original = x;
-        long long int reverse = 0;
+        long int original = x;
+        long int reverse = 0;
 
         while (x > 0) {
-            long long int digit = x % 10;
+            long int digit = x % 10;
             reverse = reverse * 10 + digit;
             x = x / 10;
         }
